Igel 2.4.0 (#168)
- implement singlular extensions
- detect drop of score and using time more aggressively
- remove IID and simplify search
- do not tt cutoff on pv moves
- simplification for nullmove pruning
- simplification for connected pawn evalution
- basic endgame knowledge to detect draw in eval/search
- retune everything + decrease number of tunnable parameters from 2600 to 1500
- apply LMR when at least one legal move has been played
- decrease reductions in LMR for pv moves

os=linux
hash=256
tc=all/60+0.6
Score of Igel 2.4.0 64 POPCNT vs Igel 2.3.1 64 POPCNT: 266 - 170 - 564 [0.548] 1000
Elo difference: 33.46 +/- 14.19

+++ src/bitboards.h
@@ -108,7 +108,7 @@ inline FLD PopLSB(U64& b)
 inline unsigned int countBits(U64 b)
 {
 #if _WIN32 || _WIN64
-    return __popcnt64(b);
+    return static_cast<unsigned int>(__popcnt64(b));
 #else
     return __builtin_popcountl(b);
 #endif
